Initial separation of system time and ntpd
